leaderboard functionality

+++ Record.cs
@@ -3,9 +3,9 @@ namespace reflexor;
 class Record
 {
     public string Username { get; set; }
-    public decimal Seconds { get; set; }
+    public double Seconds { get; set; }
 
-    public Record(string username, decimal seconds)
+    public Record(string username, double seconds)
     {
         Username = username;
         Seconds = seconds;
